Refactor models and add Hive serialization
