+++ lib/features/pages/home/data/models/doctor_speciality_response_model.dart
@@ -1,0 +1,24 @@
+import 'package:json_annotation/json_annotation.dart';
+
+import 'doctor_speciality_model.dart';
+
+part 'doctor_speciality_response_model.g.dart';
+
+@JsonSerializable()
+class DoctorSpecialityResponseModel {
+  final String message;
+  @JsonKey(name: "data")
+  final List<DoctorSpecialityModel> homeData;
+  final bool status;
+  final int code;
+
+  DoctorSpecialityResponseModel({
+    required this.message,
+    required this.homeData,
+    required this.status,
+    required this.code,
+  });
+
+  factory DoctorSpecialityResponseModel.fromJson(Map<String, dynamic> json) =>
+      _$DoctorSpecialityResponseModelFromJson(json);
+}
